# docs/swagger.yaml
basePath: /api
definitions:
  dto.UserLoginOutput:
    properties:
      access_token:
        description: access_token
        type: string
      expires_at:
        description: expires_in
        type: string
    type: object
host: :8080
info:
  contact:
    email: YuanShengSH@outlook.com
    name: Cheney
    url: http://adaday.cn
  description: 接口描述
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ginShoppingMall API
  version: "1.0"
paths:
...
  /api/user/login:
    post:
      description: 用户登录
      parameters:
      - description: 用户名
        in: path
        name: name
        required: true
        type: string
      - description: 密码
        in: path
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserLoginOutput'
      summary: 用户登录
      tags:
      - API USER 用户
  /api/user/register:
    post:
      description: 用户注册
      parameters:
      - description: 手机号
        in: path
        name: mobile
        required: true
        type: integer
      - description: 用户名
        in: path
        name: name
        required: true
        type: string
      - description: 密码
        in: path
        name: password
        required: true
        type: string
      - description: 手机验证码
        in: path
        name: captcha
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - API USER 用户
swagger: "2.0"